+++ promise.mli
@@ -1,4 +1,6 @@
-type +'a t = private < > Js_of_ocaml.Js.t
+type +'a promise
+
+type +'a t = 'a promise Js_of_ocaml.Js.t
 (** The Promise object represents the eventual completion (or failure) of an asynchronous
 	operation, and its resulting value. *)